classification/main_dashboard/requirements.txt
streamlit == 0.75.0
matplotlib == 3.3.4
numpy == 1.19.5
Pillow == 8.1.0
tensorflow == 2.4.0
requests == 2.25.1